+++ db/migrate/20240329165346_create_socializer_comments.socializer.rb
@@ -4,8 +4,8 @@
 class CreateSocializerComments < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_comments do |t|
-      t.integer  :author_id, null: false
-      t.text     :content,   null: false
+      t.bigint :author_id, null: false
+      t.text :content, null: false
 
       t.timestamps
     end
