+++ db/migrate/20240329165349_create_socializer_ties.socializer.rb
@@ -4,7 +4,7 @@
 class CreateSocializerTies < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_ties do |t|
-      t.integer :contact_id, index: true, null: false
+      t.bigint :contact_id, index: true, null: false
       t.references :circle, null: false
 
       t.timestamps
